requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
python-dotenv==1.0.1
openai==1.37.1
httpx==0.27.2
python-multipart==0.0.9
